GUI for LEGO MINDSTORMS EV3
Still in the process of adding and building features to the GUI such as sending a map and receiving coordinates from the robot. However, the user can now create a map by adding x- and y-coordinates to the text fields and the GUI will draw walls to the canvas. It can also make a connection to the robot.