Observer
This design pattern is written in Java and it demonstrates on how Observable and Observer interact with each other. When Observable changes its state, it will automatically notify all Observers about it, even though Observable is not aware of how many of them exist. Using with this clock example, it will update the current time every second.