Adapter
The purpose of this exercise is to show example on how two incompatible classes with different features will work together with adapter without modifying their source code. In this exercise I will use PS3 and PS4 consoles as an example and with SuperAdapter (that uses existing interface of Pelikonsoli) it is possible to play PS3 games on PS4 console.